engine: move LocalTargetBuildAndDeployer into buildcontrol
