+++ internal/cli/wire_gen.go
@@ -258,7 +258,7 @@ func wireCmdUp(ctx context.Context, analytics3 *analytics.TiltAnalytics, cmdTags
 	dockerComposeClient := dockercompose.NewDockerComposeClient(localEnv)
 	imageBuilder := buildcontrol.NewImageBuilder(dockerBuilder, execCustomBuilder, updateMode)
 	dockerComposeBuildAndDeployer := buildcontrol.NewDockerComposeBuildAndDeployer(dockerComposeClient, switchCli, imageBuilder, clock)
-	localTargetBuildAndDeployer := engine.NewLocalTargetBuildAndDeployer(clock)
+	localTargetBuildAndDeployer := buildcontrol.NewLocalTargetBuildAndDeployer(clock)
 	buildOrder := engine.DefaultBuildOrder(liveUpdateBuildAndDeployer, imageBuildAndDeployer, dockerComposeBuildAndDeployer, localTargetBuildAndDeployer, updateMode, env, runtime)
 	spanCollector := tracer.NewSpanCollector(ctx)
 	traceTracer, err := tracer.InitOpenTelemetry(ctx, spanCollector)
@@ -430,7 +430,7 @@ func wireCmdCI(ctx context.Context, analytics3 *analytics.TiltAnalytics, subcomm
 	dockerComposeClient := dockercompose.NewDockerComposeClient(localEnv)
 	imageBuilder := buildcontrol.NewImageBuilder(dockerBuilder, execCustomBuilder, updateMode)
 	dockerComposeBuildAndDeployer := buildcontrol.NewDockerComposeBuildAndDeployer(dockerComposeClient, switchCli, imageBuilder, clock)
-	localTargetBuildAndDeployer := engine.NewLocalTargetBuildAndDeployer(clock)
+	localTargetBuildAndDeployer := buildcontrol.NewLocalTargetBuildAndDeployer(clock)
 	buildOrder := engine.DefaultBuildOrder(liveUpdateBuildAndDeployer, imageBuildAndDeployer, dockerComposeBuildAndDeployer, localTargetBuildAndDeployer, updateMode, env, runtime)
 	spanCollector := tracer.NewSpanCollector(ctx)
 	traceTracer, err := tracer.InitOpenTelemetry(ctx, spanCollector)
